move patch to the utils package
